+++ Web Ui/src/application/views/assignmentsViews/useCases/createAssingmentsAdapter.ts
@@ -1,5 +1,5 @@
 import { AssignmentDataObject } from "../../../../modules/Assignments/domain/assignmentInterfaces";
-import { createAssignment } from "../../../../repositories/assignment.API";
+import { createAssignment } from "../../../../modules/Assignments/repository/assignment.API";
 
 export const createAssignmentsUseCase = async (assignmentData: AssignmentDataObject) => {
     return await createAssignment(assignmentData);
