+++ Web Ui/src/application/views/assignmentsViews/useCases/fetchAssignmentsApater.ts
@@ -1,4 +1,4 @@
-import { fetchAssignments } from "../../../../repositories/assignment.API";
+import { fetchAssignments } from "../../../../modules/Assignments/repository/assignment.API";
 
 export const fetchAssignmentsUseCase = async () => {
   return await fetchAssignments();
